requirements.txt
streamlit
Pillow
google-generativeai
matplotlib
reportlab 